+++ Makefile
@@ -3,7 +3,7 @@
 CFLAGS += -g -Wall
 EXECUTABLE = sensord sensorcal compdata
 _OBJ = ms5611.o ams5915.o ads1110.o main.o nmea.o timer.o KalmanFilter1d.o cmdline_parser.o configfile_parser.o vario.o AirDensity.o 24c16.o
-_OBJ_CAL = 24c16.o ams5915.o sensorcal.o
+_OBJ_CAL = 24c16.o ams5915.o sensorcal.o ms5611.o
 _OBJ_COMPDATA = ms5611.o compdata.o timer.o cmdline_parser.o configfile_parser.o
 OBJ = $(patsubst %,$(ODIR)/%,$(_OBJ))
 OBJ_CAL = $(patsubst %,$(ODIR)/%,$(_OBJ_CAL))
@@ -19,7 +19,7 @@ $(ODIR)/%.o: %.c
 	mkdir -p $(ODIR)
 	$(CC) -DVERSION_GIT=\"$(GIT_VERSION)\" -c -o $@ $< $(CFLAGS)
 	
-all: sensord cal compdata
+all: sensord sensorcal compdata
 
 version.h: 
 	@echo Git version $(GIT_VERSION)
@@ -31,7 +31,7 @@ doc:
 sensord: $(OBJ)
 	$(CC) -g -o $@ $^ $(LIBS)
 	
-cal: $(OBJ_CAL)
+sensorcal: $(OBJ_CAL)
 	$(CC) -g -o $@ $^
 
 compdata: $(OBJ_COMPDATA)
