+++ README
@@ -1,0 +1,32 @@
+Release Notes: 
+Jan 5 2021
+
+1) It has been found that timing jitter on the MS5611 pressure sensors results in glitches.  As a result, an attempt has been made
+to reduce jitter, detect jitter, and to compensate for jitter when detected.  This compensation scheme requires the introduction
+of a new program called "compdata".  Typical usage is:
+
+/opt/bin/compdata -s -c /opt/conf/sensord.conf
+
+The program will prompt you to enter the number of data points (in thousands), up to 9999.  It should be noted that it takes roughly
+1 minute per thousand data points.  Recommendation is you choose a minimum of 5.  This can be entered either with a keyboard or 
+by use of the joystick or rotary encoders in an intuitive fashion.
+
+Upon completion of the program the configuration file (/opt/conf/sensord.conf) will have compensation data appended to it, which will
+automatically be used by sensord the next time it starts.  The configuration file should already contain default compensation data.
+Whenever the compdata is run and data is saved, it will append to the end of the config file.  Sensord will always choose the
+compensation data that appears last in the file.
+
+PLEASE NOTE that compdata will report whether it thinks values it determines are good or bad.  At this point it may be best to
+disregard "BAD" if found describing coefficients.  However, Overrun/underrun errors, or > 70 on std deviation is bad.  If these
+occur, you should re-run compdata.
+
+2) Another new change is in the cal program.  Cal is used to determine the mean offset on the pitot sensor when there is no pressure 
+differential.  Previously, cal took 10 measurements, ~1 second apart, reported them and averaged them.  The data is then stored on 
+the EEPROM.  The new version will take 800 measures 12.5ms apart, average them, and store the data to the EEPROM, but will not report
+them.  This should give more accurate results.
+
+3) At this point in time it is recommended that sensord, variod, and pulseaudio are run as a forked service or from the commandline.
+When run as a regular service this can make the timing jitter worse, and results in erroneous vario readings, as well as clicks and 
+pops on the audio.  It is not presently understood why this makes a difference.
+
+-------------------------------------------------------------------------------------------------------------------------------------
